requirements.txt
llama-index
llama-parse
mistralai
python-dotenv
ipykernel # if using Jupyter
openai 
azure-ai-documentintelligence 
azure-identity
azure-storage-blob
boto3
langchain-community
langchain-unstructured